fixup! refactor: Move meta variable name to module where it is used
