+++ pkg/handlers/metadata.go
@@ -7,4 +7,7 @@ import "github.com/d2iq-labs/capi-runtime-extensions/api/v1alpha1"
 
 const (
 	MetadataDomain = v1alpha1.APIGroup
+
+	// MetaVariableName is the meta cluster config patch variable name.
+	MetaVariableName = "clusterConfig"
 )
